# internal/segments/segment_handlers/create.go
package segment_handlers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"avito-internship-2023/internal/pkg/common"
	"avito-internship-2023/internal/segments"

	"github.com/gin-gonic/gin"
	"github.com/go-playground/validator/v10"
)

type creator interface {
	CreateSegment(dto segments.CreateSegmentDTO) error
}

type CreateHandler struct {
	creator  creator
	validate *validator.Validate
}

func NewCreateHandler(creator creator, validate *validator.Validate) *CreateHandler {
	return &CreateHandler{
		creator:  creator,
		validate: validate,
	}
}

// Handle of segment_handlers/CreateHandler
// @Tags segment
// @Description Validates given segment creates it
// @Accept json
// @Param input body segments.CreateSegmentDTO true "slug and probability for each user to be added to segment"
// @Success 204
// @Router /segments/create [post]
func (handler *CreateHandler) Handle(c *gin.Context) {
	var dto segments.CreateSegmentDTO
	if err := json.NewDecoder(c.Request.Body).Decode(&dto); err != nil {
		err = fmt.Errorf("%w: %w", common.ErrJSONUnmarshalling, err)
		_ = c.Error(err)
		return
	}

	if err := handler.validate.Struct(dto); err != nil {
		_ = c.Error(err)
		return
	}

	if err := handler.creator.CreateSegment(dto); err != nil {
		_ = c.Error(err)
		return
	}

	c.Status(http.StatusNoContent)
}